+++ views/reports.percentiles.php
@@ -144,7 +144,10 @@ $html_page = (new CHtmlPage())->setTitle($title);
 
 $controls = new CList();
 
-foreach (['json' => _('Export JSON'), 'csv' => _('Export CSV'), 'yaml' => _('Export YAML')] as $format => $label) {
+foreach (
+	['json' => _('Export JSON'), 'csv' => _('Export CSV'), 'yaml' => _('Export YAML'), 'pdf' => _('Export PDF')]
+	as $format => $label
+) {
 	$export_url = (new CUrl('zabbix.php'))
 		->setArgument('action', 'morereporting.percentiles.'.$format)
 		->setArgument('filter_groupids', $data['filter']['groupids'])